+++ refena/lib/src/notifier/rebuildable.dart
@@ -1,5 +1,6 @@
 import 'package:meta/meta.dart';
 import 'package:refena/src/action/redux_action.dart';
+import 'package:refena/src/notifier/base_notifier.dart';
 import 'package:refena/src/observer/event.dart';
 import 'package:refena/src/reference.dart';
 
@@ -7,7 +8,7 @@ import 'package:refena/src/reference.dart';
 /// It might be a [Widget], a [ViewProvider] or a [WatchAction].
 /// A [Ref] holds a [Rebuildable] to make [ref.watch] work.
 @internal
-abstract class Rebuildable implements LabeledReference {
+abstract interface class Rebuildable implements LabeledReference {
   /// Schedule a rebuild (in the next frame).
   /// Optionally pass the [changeEvent], or [rebuildEvent]
   /// that triggered the rebuild.
@@ -17,6 +18,16 @@ abstract class Rebuildable implements LabeledReference {
 
   /// Whether this [Rebuildable] is disposed and should be removed.
   bool get disposed;
+
+  /// Only for [ElementRebuildable]. Noop for others.
+  /// Allows for further cleanup logic.
+  void onDisposeWidget();
+
+  /// Only for [ElementRebuildable]. Noop for others.
+  /// Notifies that a new [BaseNotifier] is being listened.
+  /// This should be called within a build method so it can unwatch
+  /// old notifiers in the next microtask.
+  void notifyListenerTarget(BaseNotifier notifier);
 
   /// A debug label for this [Rebuildable].
   @override