+++ refena_flutter/lib/src/scope.dart
@@ -185,7 +185,7 @@ class RefenaScope extends StatefulWidget implements RefenaContainer {
   }
 
   @override
-  Dispatcher<N, T> redux<N extends BaseReduxNotifier<T>, T, E extends Object>(
+  Dispatcher<N, T> redux<N extends BaseReduxNotifier<T>, T>(
     ReduxProvider<N, T> provider,
   ) {
     return Dispatcher(